Set 2: Challenge 10, implement CBC done
Ended up being pretty gnarly to use the openssl lib and have it _not_
pad by default since I was implementing it. Due to this, I also needed
to implement detecting and removing the padding on the output.

Turns out that is Challenge 15 so that's kind of done too, but I'll add
a test for their examples when I get there.