Set 3: Challenge 20
Hacky implementation but does the trick. Pretty clever to treat nonce
reuse as a (which it is) repeating xor so that transposing the first
byte of everything together ends up a set of uniformly selected letters
across the plaintext that are all xored with the same value.

So we solves that as a single byte xor, brute force + character analysis
and can solve for the key stream one byte at a time and then decrypt.