Add initial provisions proof of assets
First pass at generating the proof of assets proof. Currently it uses an
H that it should not be but will work for now for example purposes. It's
easy to swap in a new one later.

The only thing left to produce for proof of assets is the pedersen
commitment to x_i * s_i where x_i is the private key. After that comes
implementation of the interactive protocol to verify the proof.