Finish API improvements to point operations
Finally get the point to have context about the curve to simplify
defining the operations without having to pass in the curve or always
have a reference to the curve the operations are performed on.

This maps way better to the math itself and will be easier to generally
comprehend and work with.

Barring a few extra clones I haven't determined how to work around this
is much nicer.