Remove some unused code and add some verification
Always verify the point's we initialize are on the curve. This ended up
breaking SEC since I picked random values but I'll fix those later.

This helps feel more confident about there not being mistakes in the ECC
operations with some cost of performance (but it's not fast now anyways
so eh)